Changed the math engine to use _mathjs if available, it now requires data paths to be inside []
